pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repo-aura"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.32.0",
    "plotly>=5.20.0",
    "psycopg2-binary>=2.9.9",
    "PyGithub>=2.3.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["repo_aura"]
